requirements.txt
aiohappyeyeballs==2.5.0
aiohttp==3.11.13
aiosignal==1.3.2
attrs==25.1.0
croniter==6.0.0
discord-py-interactions==5.14.0
discord-typings==0.9.0
dynaconf==3.2.10
emoji==2.14.1
frozenlist==1.5.0
idna==3.10
loguru==0.7.3
multidict==6.1.0
propcache==0.3.0
python-dateutil==2.9.0.post0
pytz==2025.1
six==1.17.0
tomli==2.2.1
typing_extensions==4.12.2
Wand==0.6.13
yarl==1.18.3